requirements.txt
accelerate==0.34.2
amqp==5.2.0
audioread==3.0.1
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
diffusers==0.30.3
filelock==3.16.1
flake8==7.1.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
fsspec==2024.9.0
greenlet==3.1.1
huggingface-hub==0.25.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
kombu==5.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
psutil==6.0.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
SQLAlchemy==2.0.35
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
zipp==3.20.2
